log-analytics-system/requirements.txt
# Python requirements for the log analytics system

# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Kafka
kafka-python==2.0.2

# Elasticsearch
elasticsearch==8.10.0

# Async HTTP client
aiohttp==3.9.0
httpx==0.25.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

# Logging and monitoring
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
